requirements.txt
django==4.1
markdown==3.4.3
django-filter==23.1
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
drf-extensions==0.7.1